pyproject.toml
[tool.poetry]
name = "dystematic-code-challenge"
version = "0.1.0"
description = ""
authors = ["Nuno Diogo da Silva <diogosilva.nuno@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Django = "^2.2"
sentry-sdk = "^0.14.1"
django-environ = "^0.4.5"
psycopg2 = "^2.8.4"
djangorestframework = "*"

[tool.poetry.dev-dependencies]
django-debug-toolbar = "*"
pytest-django = "*"
pytest-cov = "*"
mixer = "*"